+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v2.1.15
 
+- New Feature - Templated Polygon Sizing [[#354](https://github.com/grafana/grafana-polystat-panel/issues/354)]
 - Force uplot dependency to resolve to 1.6.32 (package is not used)
 - Fixes [[#411](https://github.com/grafana/grafana-polystat-panel/issues/411)]
 - Fixes [[#404](https://github.com/grafana/grafana-polystat-panel/issues/404)]